libgambatte/cartridge: allow mapping rom bank0 in more cases/please blargg
blargg seems to be of the opinion that mbc1 allows mapping bank0 via overflow
for carts with few rom banks if his tests are anything to go by. I have no way
of verifying this, but am changing it because it confuses people when his tests
loop forever after succeeding.