libgambatte: fix lfsr reg init value
all bits should be high, not just the lower 8.

most likely this is a remnant from the initial implementation
based on GBSOUND.txt which has what is most likely a brain fart
0xFF value for this considering the rest of the description
is inconsistent whith that.

belmont's revenge (which has some great tracks) is affected by
this, so this one is a pretty big deal to me. konami collection
vol 4 has a pretty sweet color version of it which I recommend
checking out.