libgambatte: push return address after pc mod.
unsurprising that the call target is fetched prior to
pushing the return address (sufficiently so to verify
this based on wtf-factor alone, really). more surprising
that the 4 cycle delay that accompanies pc mods also
precedes it.

"woody woodpecker racing" barely avoids an oam dma bus
conflict due to this fact.